HoleInString.c..Finding numbers of holes in a string
The first line contains a single integer t <= 40, the number of test cases. t test cases follow. The only line of each test case contains a non-empty text composed only of uppercase letters of English alphabet. The length of the text is less then 100. There are no any spaces in the input